[*] move Logger to appropriate file
